Generalized packet-parsing subsystem to allow for much more general parsing behavior.
To allow for more complex parsing of received packet data, parsing rules were
generalized into functions associated with fields on a particular response
packet. These parsing functions are expected to accept an instance of an XBee
or ZigBee object, and the partially-parsed contents of the received data
packet (the parsing rules are iterated over, so any parsing rule after the
first will see the resulting modification applied by the first rule).

After processing the data, the function must return the new data to store
in the field of the indicated name. This will replace the raw data in that
field.














^G Get Help  ^O WriteOut  ^R Read File ^Y Prev Page ^K Cut Text  ^C Cur Pos
^X Exit      ^J Justify   ^W Where Is  ^V Next Page ^U UnCut Text^T To Spell